setup role and create some file admin html

+++ SpringThesis/src/main/java/com/ts/controllers/api/ApiBoardMemberController.java
@@ -2,8 +2,9 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package com.ts.controllers;
+package com.ts.controllers.api;
 
+//import com.ts.controllers.*;
 import com.ts.pojo.Board;
 import com.ts.pojo.BoardMember;
 import com.ts.pojo.BoardMemberPK;
@@ -14,6 +15,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import com.ts.services.UsersService;
+import jakarta.servlet.http.HttpServletRequest;
+import java.util.List;
 import java.util.Map;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -35,8 +38,13 @@ public class ApiBoardMemberController {
     @Autowired
     private BoardMemberService boardMemberService;
 
-    @PostMapping("/board-members")
-    public ResponseEntity<?> addBoardMember(@RequestBody Map<String, String> payload) {
+    @PostMapping("/secure/board-members")
+    public ResponseEntity<?> addBoardMember(@RequestBody Map<String, String> payload, HttpServletRequest request) {
+        String role = (String) request.getAttribute("role");
+        if (!"ROLE_ADMIN".equals(role) && !"ROLE_MINISTRY".equals(role)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body(Map.of("error", "Bạn không có quyền."));
+        }
         try {
             BoardMember member = this.boardMemberService.add(payload);
             return new ResponseEntity<>(member, HttpStatus.CREATED);
@@ -53,7 +61,12 @@ public class ApiBoardMemberController {
     @GetMapping("/board-members/board_id/{id}")
     public ResponseEntity<?> listMemberOfBoardById(@PathVariable("id") int boardId) {
         try {
-            return new ResponseEntity<>(this.boardMemberService.getBoardMembersByBoardId(boardId), HttpStatus.OK);
+            List<BoardMember> bm = this.boardMemberService.getBoardMembersByBoardId(boardId);
+            if (bm == null) {
+                return ResponseEntity.notFound().build();
+            } else {
+                return new ResponseEntity<>(bm, HttpStatus.OK);
+            }
         } catch (NumberFormatException ex) {
             return ResponseEntity.badRequest().body(Map.of("error", "ID không hợp lệ"));
         } catch (IllegalArgumentException ex) {
@@ -65,9 +78,14 @@ public class ApiBoardMemberController {
     }
 
     // DELETE /api/board-members?boardId=1&lecturerId=2
-    @DeleteMapping("board-members")
+    @DeleteMapping("/secure/board-members")
     public ResponseEntity<?> deleteBoardMember(@RequestParam("boardId") int boardId,
-            @RequestParam("lecturerId") int lecturerId) {
+            @RequestParam("lecturerId") int lecturerId, HttpServletRequest request) {
+        String role = (String) request.getAttribute("role");
+        if (!"ROLE_ADMIN".equals(role) && !"ROLE_MINISTRY".equals(role)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body(Map.of("error", "Bạn không có quyền."));
+        }
         try {
             boardMemberService.removeBoardMember(boardId, lecturerId);
             return ResponseEntity.noContent().build();
